[issue#78/issue#79] Pagination + Sorting POC [issue#78/issue#79] Pagination + Sorting POC
**Why**
We want to be able to paginate and sort our backend endpoint to allow efficient queries to the database as the number of records grow.

**How**
Add pagination and sorting to the league end point. Created new field validator that can check if the passed in field is a valid field. Updated yml files due to errors on schema imports.